index and doodle htmls added
Index.html is the entry point. At this point, when there is some text and chromatify button is pressed, a random color and shape from colors and shapes arrays are stored in local storage. The doodle HTML then rendered inside the index.html and it reads from the localStorage to update its' values. Then users can save the file as .png. 

To do: 
- Select a set of colors maintaining color harmony (grid background and cell background both)
- checkout rotation and scaling more in-depth
- Send/receive data and map probability to shapes, colors, and rotation